pyproject.toml
[tool.poetry]
name = "trisg-kt"
version = "0.1.0"
description = "TriSG-KT: Triple-Sparse Semantic Graph Knowledge Tracing"
authors = ["Yu Yin <yxonic@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
torch = "^2.3.1"
numpy = "^2.0.2"
tomlkit = "^0.14.0"
scipy = "^1.13.1"
scikit-learn = "^1.4.2"
transformers = "^4.49.0"
sentence-transformers = "^5.3.0"
matplotlib = "^3.10"
pandas = "^2.2"
pyyaml = "^6.0"
tqdm = "^4.67"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["DTransformer/tests"]
